Use UInt64 for download limit remote configs

+++ SignalServiceKit/Backups/Attachments/BackupAttachmentDownloadQueueStatusManager.swift
@@ -628,7 +628,7 @@ public class BackupAttachmentDownloadQueueStatusManagerImpl: BackupAttachmentDow
     }
 
     private nonisolated func getRequiredDiskSpace() -> UInt64 {
-        return UInt64(remoteConfigManager.currentConfig().attachmentMaxEncryptedBytes) * 5
+        return remoteConfigManager.currentConfig().attachmentMaxEncryptedBytes * 5
     }
 
     @objc
